remove VLA (fix `-Wvla`).
VLA are not supported in C++, that's a C99 feature not a C++ one (and
even in C, it becames optional since C11).

Even if they were supported, it's not a good idea to use them
(especially with an unbounded size such as `this->n`) this is a recipe
for easy stack overflows.

Now, dumbly replacing VLAs by `std::vector` simply won't do (the code
becomes 50% slower). But, we can implement a scratch space inside the
`gf::NF4` object that can replaces the VLAs and get a nice 10% speedup :)